register terminado
