+++ src/App.jsx
@@ -5,7 +5,7 @@ import Home from './pages/Home/Home'
 import Movies from './pages/Movies/Movies'
 import MoviesDetail from './pages/MoviesDetail/MoviesDetail'
 import Series from './pages/Series/Series'
-import DetailSeries from './components/Seriedetail/Seriedetail'
+import SerieDetail from './pages/Seriedetail/Seriedetail'
 import Register from './pages/Register/register'
 
 
@@ -22,7 +22,7 @@ const App = () => {
 	<Route exact path='/register' element={<Register/>} />
 	<Route exact path='/series' element={<Series/>} />
 	<Route exact path='/moviesdetail/:id' element={<MoviesDetail/>} />
-	<Route exact path='/detailSeries' element={<DetailSeries/>}/>
+	<Route exact path='/detailSeries/:id' element={<SerieDetail/>}/>
  </Routes>
 </BrowserRouter>
 )
