+++ src/components/CardSerie/card.module.css
@@ -1,0 +1,12 @@
+.containerMax{
+    width: 216px;
+    height: 276px;
+    background-size: cover;
+    margin: 20px 10px;
+}
+
+img{
+    width: 216px;
+    height: 276px;
+    background-size: cover
+}
