--- src/components/Seriedetail/seriedetail.css
@@ -1,125 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Roboto&display=swap');
-
-.sectionDetailSerie {
-    background-color: rgb(34, 32, 32);
-    overflow: auto;
-}
-
-.nameSeie {
-    display: flex;
-    justify-content: left;
-    margin-left: 30px;
-    font-family: 'Roboto', sans-serif;
-}
-
-.containerDivDetail {
-    display: flex;
-    margin-left: 30px;
-}
-
-.imgDetailSerie {
-    width: 300px;
-    height: 340px;
-    margin-right: 30px;
-}
-
-.nameDescripcion {
-    margin-top: 0;
-    color: #2980B9;
-    margin-bottom: 1px;
-    font-size: 25px;
-    font-family: 'Roboto', sans-serif;
-}
-
-.descripcionSerie {
-    color: #fff;
-    margin-top: 1px;
-    font-size: 18px;
-    font-family: 'Roboto', sans-serif;
-}
-
-.nameRaiting {
-    color: #2980B9;
-    margin-bottom: 1px;
-    font-size: 25px;
-    font-family: 'Roboto', sans-serif;
-}
-
-.divEstrellas {
-    display: flex;
-    cursor: pointer;
-}
-
-.nameActores {
-    color: #2980B9;
-    margin-bottom: 1px;
-    font-size: 25px;
-    font-family: 'Roboto', sans-serif;
-}
-
-.actoresSeries {
-    color: #fff;
-    margin-top: 1px;
-    font-size: 18px;
-    font-family: 'Roboto', sans-serif;
-}
-
-.nameDirector {
-    color: #E01C1C;
-    margin-bottom: 1px;
-    font-size: 25px;
-    font-family: 'Roboto', sans-serif;
-}
-
-.directorSerie {
-    color: #fff;
-    margin-top: 1px;
-    font-size: 18px;
-    font-family: 'Roboto', sans-serif;
-}
-
-.divButtonCompraSerie {
-    display: flex;
-    justify-content: center;
-}
-
-.buttonCompraSerie {
-    margin-top: 30px;
-    padding: 13px 400px;
-    color: #fff;
-    background-color: #2980B9;
-    border: none;
-    border-radius: 5px;
-    font-size: 30px;
-    cursor: pointer;
-}
-
-.divContainerTC {
-    display: flex;
-    justify-content: center;
-}
-
-.divTemp {
-    margin-top: 60px;
-    margin-right: 300px;
-}
-
-.divCap {
-    margin-top: 60px;
-    margin-left: 300px;
-}
-
-.p {
-    margin-bottom: 0;
-    font-size: 30px;
-    font-family: 'Roboto', sans-serif;
-}
-
-.selectDetail {
-    margin-bottom: 80px;
-    padding: 6px 30px;
-}
-
-.divVideo {
-    margin: 10px 200px 200px 200px;
-}
