+++ src/pages/Series/Series.jsx
@@ -1,139 +1,50 @@
 import React from 'react'
-import { NavLink } from 'react-router-dom'
-import './series.css'
+import data from '../../data'
+import style from './series.module.css'
+import Card from '../../components/CardSerie/Card'
 
 const Series = () => {
+
+  const series = data
+
   return (
-
-    <section className="section">
-      
-      <p className="pTitle">Todas la series</p>
-
-      <div className="containerFilter">
-      
-        <div className="div">
-          
-          <p className="p">Categoria</p>
-          
-          <select className="select">
-            <option>Filtrar</option>
+    <section className={style.containerMax}>
+      <h1>Todas las series</h1>
+      <div className={style.filters}>
+        <div>
+          <span>Categoría</span>
+          <select>
+            <option>Select categoria</option>
           </select>
-          
         </div>
-          
-        <div className="div">
-        
-          <p className="p">Año</p>
-        
-          <select className="select">
-            <option>Filtrar</option>
+        <div>
+          <span>Año</span>
+          <select>
+            <option>Select año</option>
           </select>
-        
         </div>
-
-        <div className="div">
-        
-          <p className="p">Puntuación</p>
-        
-          <select className="select">
-            <option>Filtrar</option>
+        <div>
+          <span>Puntuación</span>
+          <select>
+            <option>Select Puntuación</option>
           </select>
-        
         </div>
-
-        <div className="div">
-        
-          <p className="p">Ordenamiento</p>
-        
-          <select className="select">
-            <option>Filtrar</option>
+        <div>
+          <span>Ordenamiento</span>
+          <select>
+            <option>select Ordenamiento</option>
           </select>
-        
         </div>
-
       </div>
-
-      <section className="containerSeries">
-        
-        <div className="divContainerCard">
-
-          <NavLink to='/detailSeries'>
-            <img className="imgSeries" src="https://d500.epimg.net/cincodias/imagenes/2020/12/31/lifestyle/1609408585_467254_1609408795_noticia_normal.jpg" alt="" />
-          </NavLink>
-
-          <NavLink to='/detailSeries'>
-            <img className="imgSeries" src="https://i.postimg.cc/dQfgLrfV/IMG-20221231-124201.jpg" alt="" />
-          </NavLink>
-
-          <NavLink to='/detailSeries'>
-            <img className="imgSeries" src="https://i.blogs.es/522d7e/mejoresfantasia2023/1366_2000.jpeg" alt="" /> 
-          </NavLink>
-
-          <NavLink to='/detailSeries'>
-            <img className="imgSeries" src="https://d500.epimg.net/cincodias/imagenes/2020/12/31/lifestyle/1609408585_467254_1609408795_noticia_normal.jpg" alt="" />
-          </NavLink>
-
-          <NavLink to='/detailSeries'>
-            <img className="imgSeries" src="https://i.postimg.cc/dQfgLrfV/IMG-20221231-124201.jpg" alt="" />
-          </NavLink>
-
-          <NavLink to='/detailSeries'>
-            <img className="imgSeries" src="https://d500.epimg.net/cincodias/imagenes/2020/12/31/lifestyle/1609408585_467254_1609408795_noticia_normal.jpg" alt="" />
-          </NavLink>
-
-          <NavLink to='/detailSeries'>
-            <img className="imgSeries" src="https://i.blogs.es/522d7e/mejoresfantasia2023/1366_2000.jpeg" alt="" />
-          </NavLink>
-
-          <NavLink to='/detailSeries'>
-            <img className="imgSeries" src="https://d500.epimg.net/cincodias/imagenes/2020/12/31/lifestyle/1609408585_467254_1609408795_noticia_normal.jpg" alt="" />
-          </NavLink>
-
-          <NavLink to='/detailSeries'>
-            <img className="imgSeries" src="https://i.blogs.es/522d7e/mejoresfantasia2023/1366_2000.jpeg" alt="" />
-          </NavLink>
-
-          <NavLink to='/detailSeries'>
-            <img className="imgSeries" src="https://d500.epimg.net/cincodias/imagenes/2020/12/31/lifestyle/1609408585_467254_1609408795_noticia_normal.jpg" alt="" />
-          </NavLink>
-
-
-          <NavLink to='/detailSeries'>
-            <img className="imgSeries" src="https://i.blogs.es/522d7e/mejoresfantasia2023/1366_2000.jpeg" alt="" />
-          </NavLink>
-
-          <NavLink to='/detailSeries'>
-            <img className="imgSeries" src="https://d500.epimg.net/cincodias/imagenes/2020/12/31/lifestyle/1609408585_467254_1609408795_noticia_normal.jpg" alt="" />
-          </NavLink>
-
-          <NavLink to='/detailSeries'>
-            <img className="imgSeries" src="https://i.blogs.es/522d7e/mejoresfantasia2023/1366_2000.jpeg" alt="" />
-          </NavLink>
-
-          <NavLink to='/detailSeries'>
-            <img className="imgSeries" src="https://d500.epimg.net/cincodias/imagenes/2020/12/31/lifestyle/1609408585_467254_1609408795_noticia_normal.jpg" alt="" />
-          </NavLink>
-
-          <NavLink to='/detailSeries'>
-            <img className="imgSeries" src="https://i.blogs.es/522d7e/mejoresfantasia2023/1366_2000.jpeg" alt="" />
-          </NavLink>
-        
+      <div className={style.Container}>
+        <div className={style.serieContainer}>
+          {series.map(({ id, image }) => (
+          <Card key={id} id={id} image={image} />
+          ))}
+        <div className={style.paginado}>paginado</div>
         </div>
-
-        <div className="divPaginado">
-
-          <button className="elementoB button">Anterior</button>
-          
-          <p className='pPaginado'>1</p>
-          <p className='pPaginado'>2</p>
-          <p className='pPaginado'>3</p>
-          
-          <button className="elementoB button">Siguiente</button>
-        
-        </div>
-        
-      </section>
-
+      </div>
+      
     </section>
   )
 }
