--- src/pages/Series/series.css
@@ -1,82 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Roboto&display=swap');
-
-.section {
-    background-color: rgb(34, 32, 32);
-    overflow: auto;
-}
-
-/* Titulo */
-
-.pTitle {
-    margin: 40px;
-    color: #fff;
-    font-size: 40px;
-    font-family: 'Roboto', sans-serif;
-}
-
-/* Select Filter */
-
-.containerFilter {
-    display: flex;
-}
-
-.div {
-    margin: 10px;
-    margin-left: 60px;
-}
-
-.select {
-    padding: 5px 20px;
-}
-
-.p {
-    color: #fff;
-    font-size: 20px;
-    font-family: 'Roboto', sans-serif;
-}
-
-/* Select series */
-
-.containerSeries {
-    margin: 60px;
-}
-
-.divContainerCard {
-    display: grid;
-    grid-template-columns: 1fr 1fr 1fr 1fr 1fr;
-    grid-gap: 20px;
-    gap: 20px;
-}
-
-.imgSeries {
-    width: 100%;
-    height: 100%;
-    cursor: pointer;
-}
-
-/* Paginado */
-
-.divPaginado {
-    display: flex;
-    justify-content: center;
-    margin-top: 50px;
-}
-
-.elementoB {
-    margin: 10px;
-}
-
-.button {
-    border: none;
-    border-radius: 3px;
-    padding: 6px;
-    font-size: 15px;
-    cursor: pointer;
-}
-
-.pPaginado {
-    color: #fff;
-    margin: 20px 20px;
-    font-size: 20px;
-    cursor: pointer;
-}
