+++ src/pages/Series/series.module.css
@@ -1,9 +1,3 @@
-body{
-    padding: 0;
-    margin: 0;
-    background-color: #101010;
-}
-
 span{
     font-size: 14px;
 }
@@ -11,7 +5,6 @@ span{
 h1{
     color: black;    
 }
-
 
 .filters{
     display: flex;
@@ -46,7 +39,7 @@ h1{
 
 }
 
-.peliculaContainer{
+.serieContainer{
     background-color: #000000;
     display: flex;
     justify-content: center;
@@ -64,4 +57,3 @@ h1{
     display: flex;
     justify-content: center;
 }
-
